refactor: Split storyboard forms page into modular component files
- Extract each form component example into separate files
- Improve code organization and maintainability
- Make individual components easier to edit and understand
- Reduce main page.tsx from ~1200 lines to ~36 lines
- Fix TypeScript issues with checkbox checked state handling

Component files created:
- input-example.tsx
- textarea-example.tsx
- select-example.tsx
- checkbox-example.tsx
- radio-group-example.tsx
- switch-example.tsx
- slider-example.tsx
- input-otp-example.tsx
- date-picker-example.tsx
